+++ src/main/java/org/asundr/RoundingRule.java
@@ -32,6 +32,6 @@ public enum RoundingRule
     NEAREST(val -> (long)Math.round(val)),
     DOWN(val -> (long)Math.floor(val)),
     UP(val -> (long)Math.ceil(val));
-    final public Function<Float, Long> method;
-    RoundingRule(final Function<Float, Long> method) { this.method = method; }
+    final public Function<Double, Long> method;
+    RoundingRule(final Function<Double, Long> method) { this.method = method; }
 }
